Fix ambiguous statement in setup readme
The phrase `Please note that you do not need to follow the rest of this guide [...]` is a bit ambiguous. My interpretation of the sentence was "You don't need to (but you can) follow the rest of this guide". Unfortunately this made me spend a few hours trying to follow the, what I thought was, optional guide. 

I just changed the wording making it obvious that the app is ready after running the command, and the rest of the guide is only to understand how it works.